chore: useSummary hook implemented
