+++ src/components/Summary/index.tsx
@@ -1,22 +1,10 @@
-import { useContext } from "react";
 import { SummaryContainer, SummaryCard } from "./styles";
 import { ArrowCircleDown, ArrowCircleUp, CurrencyDollar } from 'phosphor-react'
-import { TransactionsContext } from "../../contexts/TransactionContext";
 import { priceFormatter } from "../../utils/currencyFormatter";
+import { useSummary } from "../../hooks/useSummary";
 
 export function Summary() {
-    const { transactions } = useContext(TransactionsContext)
-
-    const summary = transactions.reduce((acc, curr) => {
-        if (curr.type === "entry") {
-            acc.entry += Number(curr.price)
-            acc.total += Number(curr.price)
-            return acc
-        }
-        acc.exit += Number(curr.price)
-        acc.total -= Number(curr.price)
-        return acc
-    }, { entry: 0, exit: 0, total: 0 })
+    const summary = useSummary()
 
     return (
         <SummaryContainer>
